requirements.txt
requests2
pandas
lxml
beautifulsoup4
selenium
